refactor(build): rename `deno_compile` rule to `deno_binary` & use macro
This change renames the `deno_compile` Buck2 rule to `deno_binary` and
converts its direct use with the usage of the `deno_binary` macro, in a
similar pattern to the `rust_binary` macro.

To help disambiguate the variable that is the full path to the `deno`
program executable, this variable is now `deno_exe` throughout the Buck2
code.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>